# Includes list correction, navigation service, and delete alert (#1)
* # include Messaging, and close window

* #Include Alert of Customer Delete

* novo

* # include navigationServices

---------

+++ Core/Features/Customers/CustomerList/CustomersListViewModel.cs
@@ -1,31 +1,47 @@
 ﻿namespace Core;
-public partial class CustomersListViewModel(ICustomerService customerService) : BaseListViewModel<Customer>
+public partial class CustomersListViewModel : BaseListViewModel<Customer>
 {
-    private readonly ICustomerService _customerService = customerService;
+    readonly ICustomerService _customerService;
 
-    public override void ApplyQueryAttributes(IDictionary<string, object> query)
+    public CustomersListViewModel(ICustomerService customerService, INavigationService navigationService) : base(navigationService)
     {
-        InitializeAsync();
+        _customerService = customerService;
+        WeakReferenceMessenger.Default.Register<MyMessage>(this, (r, m) => { RefreshDataAsync(); });
+    }
+    public override void ApplyQueryAttributes(IDictionary<string, object> query) => InitializeAsync();
+
+    [RelayCommand]
+    async Task Add()
+    {
+#if WINDOWS
+        var viewModel = new CustomerCreateViewModel(_customerService, _navigationService);
+        await _navigationService.NavigateToAsync($"{nameof(CustomerCreatePage)}", new CustomerCreatePage(viewModel),viewModel);
+#else
+        await _navigationService.NavigateToAsync($"{nameof(CustomerCreatePage)}");
+#endif
     }
 
     [RelayCommand]
-    async Task Add() => await Shell.Current.GoToAsync(nameof(CustomerCreatePage));
-
-    [RelayCommand]
     async Task Edit(Customer item)
-    {        
-        var param = new ShellNavigationQueryParameters
-        {
-            { "CustomerId", item.Id }
-        };
-        await Shell.Current.GoToAsync(nameof(CustomerCreatePage), param);        
+    {
+        var param = new ShellNavigationQueryParameters { { "CustomerId", item.Id } };
+#if WINDOWS
+        var viewModel = new CustomerCreateViewModel(_customerService, _navigationService);
+        await _navigationService.NavigateToAsync($"{nameof(CustomerCreatePage)}", new CustomerCreatePage(viewModel),viewModel, param);
+#else
+        await _navigationService.NavigateToAsync($"{nameof(CustomerCreatePage)}", routeParameters: param);
+#endif
+
     }
     [RelayCommand]
     async Task Delete(Customer item)
     {
-        _customerService.RemoveCustomer(item);
-        Items?.Clear();
-        RefreshDataAsync();
+        var result = await Shell.Current.DisplayAlert("Alerta", "Deseja mesmo deletar o cliente?", "sim", "não");
+        if (result)
+        {
+            await _customerService.RemoveCustomer(item);
+            await RefreshDataAsync();
+        }
     }
 
     protected override Task<IEnumerable<Customer>> GetDataAsync()
